requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
google-cloud-storage==2.16.0
google-cloud-bigquery==3.25.0